refactored code and added docs
